+++ cmd/blocklist/list.go
@@ -20,12 +20,12 @@ var listCmd = &cobra.Command{
 		req := apiClient.BlocklistAPI.BlocklistGet(ctx)
 
 		if cmd.Flags().Changed("take") {
-			v, _ := cmd.Flags().GetFloat32("take")
-			req = req.Take(v)
+			v, _ := cmd.Flags().GetInt("take")
+			req = req.Take(float32(v))
 		}
 		if cmd.Flags().Changed("skip") {
-			v, _ := cmd.Flags().GetFloat32("skip")
-			req = req.Skip(v)
+			v, _ := cmd.Flags().GetInt("skip")
+			req = req.Skip(float32(v))
 		}
 		if cmd.Flags().Changed("search") {
 			v, _ := cmd.Flags().GetString("search")
@@ -42,8 +42,8 @@ var listCmd = &cobra.Command{
 }
 
 func init() {
-	listCmd.Flags().Float32("take", 20, "Number of items to return")
-	listCmd.Flags().Float32("skip", 0, "Number of items to skip")
+	listCmd.Flags().Int("take", 20, "Number of items to return")
+	listCmd.Flags().Int("skip", 0, "Number of items to skip")
 	listCmd.Flags().String("search", "", "Search by title")
 	listCmd.Flags().String("filter", "", "Filter: all, manual, blocklistedTags")
 	Cmd.AddCommand(listCmd)
