+++ cmd/search/trending.go
@@ -17,14 +17,14 @@ var trendingCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		apiClient, ctx, isVerbose := newAPIClient()
 
-		page, _ := cmd.Flags().GetFloat32("page")
+		page, _ := cmd.Flags().GetInt("page")
 		language, _ := cmd.Flags().GetString("language")
 		mediaType, _ := cmd.Flags().GetString("media-type")
 		timeWindow, _ := cmd.Flags().GetString("time-window")
 
 		req := apiClient.SearchAPI.DiscoverTrendingGet(ctx)
 		if cmd.Flags().Changed("page") {
-			req = req.Page(page)
+			req = req.Page(float32(page))
 		}
 		if cmd.Flags().Changed("language") {
 			req = req.Language(language)
@@ -42,7 +42,7 @@ var trendingCmd = &cobra.Command{
 }
 
 func init() {
-	trendingCmd.Flags().Float32("page", 1, "Page number")
+	trendingCmd.Flags().Int("page", 1, "Page number")
 	trendingCmd.Flags().String("language", "en", "Language code")
 	trendingCmd.Flags().String("media-type", "all", "Media type (all, movie, tv)")
 	trendingCmd.Flags().String("time-window", "day", "Time window (day, week)")
